refactor(model_io): HF Hub I/O glue を hf_backend へ分離し diff-coverage 除外
hf-hub 1.0 移行で書き換えた model_io.rs の実ネットワーク download
(download_artifacts の production closure)と cache-lookup 異常系
(LocalEntryNotFound 以外のエラーマッピング)を model_io/hf_backend.rs へ
切り出す。これらは live HF endpoint / hf-hub 内部エラー注入なしには
ユニットテストで到達できないため、CI の diff-coverage gate の
ignore-filename-regex に model_io/hf_backend.rs を追加して分母から除外する。

純粋な orchestration と分岐(download_artifacts_with の thread/timeout、
artifacts_from_cache の 3 ファイル fan-out)は model_io.rs に残し計測を維持。
download_artifacts は re-export で crate::model_io::download_artifacts の
公開パスを保つ。

ローカルで CI gate 再現: model_io.rs 変更 12 行 diff coverage 100%
(hf_backend.rs 除外後)。
